+++ CardGameStore/Services/Implementations/NfceEmissionService.cs
@@ -405,7 +405,7 @@ public class NfceEmissionService : INfceEmissionService
 
         var pfxBytes    = Convert.FromBase64String(_enc.Decrypt(cfg.CertificadoPfxEncrypted!));
         var senha       = _enc.Decrypt(cfg.CertificadoSenhaEncrypted!);
-        var certificado = new X509Certificate2(pfxBytes, senha, X509KeyStorageFlags.Exportable);
+        var certificado = Pkcs12Loader.Abrir(pfxBytes, senha);
         var cfgCertificado = new ConfiguracaoCertificado { ArrayBytesArquivo = pfxBytes, Senha = senha };
 
         var estado   = Enum.Parse<Estado>(cfg.Uf);